dalle_telegram/requirements.txt
openai>=1.30.0
python-telegram-bot>=20.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
